Moved enemy class to separate file
